Insert at index

+++ doubly_linked_lists/7-insert_dnodeint.c
@@ -11,7 +11,7 @@
 
 dlistint_t *insert_dnodeint_at_index(dlistint_t **h, unsigned int idx, int n)
 {
-	size_t i = 0, count = 0;
+	unsigned int i = 0, count = 0;
 	dlistint_t *new_node = malloc(sizeof(dlistint_t)), *ptr, *save, *lengh = *h;
 
 	/*test if index outside the list*/
